Backend/requirements.txt
fastapi
uvicorn
torch
pydantic
python-dotenv
numpy
